Update clang to 15.0.6
I also removed clang version number strings ("12.0.0") where possible to
make similar upgrades less painful in the future.  Unfortunately,
there's a limit to what can be achieved here, because the file structure
of clang releases is not actually stable (so just replacing the version
number with a wildcard is not generally enough).

Tests don't build yet.